refactor(bayn): simplify non-database runtime
