+++ docs/bayn/contracts/v1.md
@@ -92,7 +92,8 @@ Runtime status is an operational HTTP view, not a persisted versioned evidence c
 - Dependency states are `UNKNOWN`, `AVAILABLE`, and `UNAVAILABLE`.
 - Readiness requires an active evidence envelope and every dependency to be `AVAILABLE`.
 - Data and evidence project to `UNKNOWN`, `CURRENT`, or `INVALID`.
-- Economic and qualification status come from the terminal qualification evidence.
+- `economic.verdict` contains the economic gate result and `qualification.verdict` contains the terminal qualification
+  result. Neither is an execution flag; authority remains a separate fixed field.
 - Accounting is `UNKNOWN`, `EXACT`, or `UNAVAILABLE`.
 - Authority is fixed at maximum `observe`; broker orders and capital promotion are always false.
