+++ docs/signal/adjusted-daily-publications.md
