+++ services/bayn/src/build.ts
@@ -1,15 +1,16 @@
 import { Effect, Schema } from 'effect'
 
 import { operationalError, type OperationalError } from './errors'
+import {
+  GitSourceRevisionSchema as SourceRevision,
+  ImageRepositorySchema as ImageRepository,
+  Sha256Schema as Sha256,
+} from './schemas'
 
 declare const __BAYN_BUILD_SOURCE_REVISION__: string
 declare const __BAYN_BUILD_IMAGE_REPOSITORY__: string
 declare const __BAYN_BUILD_STRATEGY_BEHAVIOR_HASH__: string
 declare const __BAYN_BUILD_STRATEGY_PARAMETER_HASH__: string
-
-const SourceRevision = Schema.String.check(Schema.isPattern(/^[a-f0-9]{40}$/))
-const ImageRepository = Schema.String.check(Schema.isPattern(/^[a-z0-9.-]+(?::[0-9]+)?\/[a-z0-9._/-]+$/))
-const Sha256 = Schema.String.check(Schema.isPattern(/^[a-f0-9]{64}$/))
 
 export const EmbeddedBuildMetadataSchema = Schema.Struct({
   sourceRevision: SourceRevision,
