+++ services/bayn/src/contracts.ts
@@ -1,38 +1,19 @@
 import { Schema } from 'effect'
 
 import { canonicalHashV1, sha256 } from './hash'
+import {
+  ImageDigestSchema as ImageDigest,
+  IsoDateSchema,
+  PositiveIntegerSchema as PositiveInteger,
+  Sha256Schema,
+  SourceRevisionSchema as SourceRevision,
+  StrictNonEmptyStringSchema as NonEmptyString,
+  SymbolSchema as SymbolName,
+  UtcInstantSchema as UtcInstant,
+  strictParseOptions as StrictParseOptions,
+} from './schemas'
 import { ContractVersion, DataFeed, DataSource, PriceAdjustment, PublicationSchema } from './types'
 
-const StrictParseOptions = { onExcessProperty: 'error' } as const
-
-type IsoDateValue = `${number}-${number}-${number}`
-
-const isIsoDate = (value: string): value is IsoDateValue => {
-  if (!/^\d{4}-\d{2}-\d{2}$/.test(value)) return false
-  const date = new Date(`${value}T00:00:00.000Z`)
-  return !Number.isNaN(date.getTime()) && date.toISOString().slice(0, 10) === value
-}
-
-const isUtcInstant = (value: string): boolean => {
-  if (!/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z$/.test(value)) return false
-  const date = new Date(value)
-  return !Number.isNaN(date.getTime()) && date.toISOString() === value
-}
-
-const NonEmptyString = Schema.String.check(
-  Schema.makeFilter((value: string) => value.length > 0 && value.trim() === value, {
-    expected: 'a non-empty string without surrounding whitespace',
-  }),
-)
-export const IsoDateSchema = Schema.String.pipe(Schema.refine(isIsoDate, { expected: 'a valid ISO date (YYYY-MM-DD)' }))
-const UtcInstant = Schema.String.check(
-  Schema.makeFilter(isUtcInstant, { expected: 'a canonical UTC instant (YYYY-MM-DDTHH:mm:ss.sssZ)' }),
-)
-export const Sha256Schema = Schema.String.check(Schema.isPattern(/^[a-f0-9]{64}$/))
-const ImageDigest = Schema.String.check(Schema.isPattern(/^sha256:[a-f0-9]{64}$/))
-const SourceRevision = Schema.String.check(Schema.isPattern(/^(?:[a-f0-9]{40}|[a-f0-9]{64})$/))
-const PositiveInteger = Schema.Int.check(Schema.isGreaterThan(0))
-const SymbolName = Schema.String.check(Schema.isPattern(/^[A-Z][A-Z0-9.-]{0,15}$/))
 const UniverseId = Schema.String.check(Schema.isPattern(/^[a-z0-9]+(?:[.-][a-z0-9]+)*$/))
 const CanonicalJson = Schema.Unknown.check(Schema.makeFilter(Schema.is(Schema.Json), { expected: 'a JSON value' }))
 
@@ -237,3 +218,5 @@ export const makeRuntimeProvenance = (input: RuntimeProvenanceInput): RuntimePro
     },
   })
 }
+
+export { IsoDateSchema, Sha256Schema } from './schemas'
