+++ services/bayn/src/qualification-audit.test.ts
@@ -3,9 +3,9 @@ import { describe, expect, test } from 'bun:test'
 import { makeEquitySeriesArtifact } from './evidence-contracts'
 import { canonicalHashV1 } from './hash'
 import { makeQualificationResult } from './qualification'
-import { auditQualification, type AuditDatabaseSnapshot, type QualificationAuditInput } from './qualification-audit'
-import { makeQualificationDossier } from './qualification-dossier'
-import { makeStrategy } from './strategy-service'
+import { auditQualification, type AuditDatabaseSnapshot, type QualificationAuditInput } from './audit/audit'
+import { makeQualificationDossier } from './audit/dossier'
+import { makeStrategy } from './strategy'
 import { summarizeEvaluation } from './risk-balanced-trend'
 import { fixtureProtocol, makeSnapshot, makeTestProvenance } from './test-fixtures'
